screenshot: update per gdk api change

+++ src/desktop/screenshot.rs
@@ -146,9 +146,9 @@ impl Into<gdk3::RGBA> for Color {
 impl Into<gdk4::RGBA> for Color {
     fn into(self) -> gdk4::RGBA {
         gdk4::RGBA {
-            red: self.red(),
-            green: self.green(),
-            blue: self.blue(),
+            red: self.red() as f32,
+            green: self.green() as f32,
+            blue: self.blue() as f32,
             alpha: 1_f32,
         }
     }
